# api/http/openapi.yaml
openapi: "3.0.3"
info:
  title: Breakfix API
  version: "1.0.0"
  description: SRE/DevOps interview practice platform
servers:
  - url: http://localhost:9090/api
    description: Local dev
paths:
...
  /challenges/{id}/content:
    get:
      operationId: getChallengeContent
      summary: Get challenge problem, solution, hints, and checkpoint metadata
      tags: [challenges]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeContent"
        "404":
          $ref: "#/components/responses/Error"
  /challenges/{id}/progress:
    get:
      operationId: getChallengeProgress
      summary: Read the controller-owned current challenge checkpoint snapshot
      tags: [environments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current checkpoint results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeProgress"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TaxonomyReference:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
    ChallengeOutcome:
      type: object
      required: [id, title, primary]
      properties:
        id:
          type: string
        title:
          type: string
        primary:
          type: boolean
    ChallengeEntrySkill:
      type: object
      required: [id, title, requires]
      properties:
        id:
          type: string
        title:
          type: string
        requires:
          type: array
          items:
            $ref: "#/components/schemas/TaxonomyReference"
    ChallengeTaxonomy:
      type: object
      required: [revision, tags, primary_outcome, outcomes, entry_skills]
      properties:
        revision:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TaxonomyReference"
        primary_outcome:
          $ref: "#/components/schemas/TaxonomyReference"
        outcomes:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeOutcome"
        entry_skills:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeEntrySkill"
    ChallengeCheckpoint:
      type: object
      required: [id, title, description]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        hint:
          type: string
        node:
          type: string
    ChallengeNode:
      type: object
      required: [name, title]
      properties:
        name:
          type: string
        title:
          type: string
    CheckpointResult:
      type: object
      required: [id, passed, summary]
      properties:
        id:
          type: string
        passed:
          type: boolean
        first_passed_at:
          type: string
          format: date-time
          nullable: true
        summary:
          type: string
        details:
          type: string
...
    ChallengeContent:
      type: object
      required: [id, title, runtime, nodes, problem, solution, hints, checkpoints, taxonomy]
      properties:
        id:
          type: string
        title:
          type: string
        runtime:
          type: string
          enum: [node, k8s]
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeNode"
        problem:
          type: string
        solution:
          type: string
        hints:
          type: object
          additionalProperties:
            type: string
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeCheckpoint"
        taxonomy:
          $ref: "#/components/schemas/ChallengeTaxonomy"
    ChallengeProgress:
      type: object
      required: [checks]
      properties:
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckpointResult"
...